Add --exclude feature to weesu
Require rewriting scan_dir so that jq would be able to filter paths
and preempt unnecessary filesystem walking. It's kind of crazy now,
but I think it works.

Also changed the default to not ignored hidden files, since they can
be ignored with the new --exclude feature and it was always
inconsistent with hidden files inside tarballs, which *aren't*
ignored.